move files around in go code
