+++ protoc-gen-twirpy/generator/template.go
@@ -1,4 +1,4 @@
-package templates
+package generator
 
 import "text/template"
