+++ protoc-gen-twirpy/go.mod
@@ -2,4 +2,7 @@ module github.com/verloop/twirpy/protoc-gen-twirpy
 
 go 1.13
 
-require github.com/golang/protobuf v1.4.2
+require (
+	github.com/golang/protobuf v1.4.2
+	google.golang.org/protobuf v1.23.0
+)
